requirements.txt
Django==2.0.7
djangorestframework==3.8.2
django-cors-headers==2.4.0
psycopg2==2.7.5
requests==2.19.1

